+++ include/monolish_blas_double.hpp
@@ -16,21 +16,55 @@ namespace monolish {
 namespace blas {
 
 void add(const vector<double> &a, const vector<double> &b, vector<double> &y);
-void add(const vector<double> &a, const double alpha, vector<double> &y);
-
 void sub(const vector<double> &a, const vector<double> &b, vector<double> &y);
-void sub(const vector<double> &a, const double alpha, vector<double> &y);
-
 void mul(const vector<double> &a, const vector<double> &b, vector<double> &y);
-void mul(const vector<double> &a, const double alpha, vector<double> &y);
-
 void div(const vector<double> &a, const vector<double> &b, vector<double> &y);
+
+void add(const vector<double> &a, const double alpha, vector<double> &y);
+void sub(const vector<double> &a, const double alpha, vector<double> &y);
+void mul(const vector<double> &a, const double alpha, vector<double> &y);
 void div(const vector<double> &a, const double alpha, vector<double> &y);
 
 void vecadd(const vector<double> &a, const vector<double> &b,
             vector<double> &y);
 void vecsub(const vector<double> &a, const vector<double> &b,
             vector<double> &y);
+
+void add(const matrix::Dense<double> &A, const matrix::Dense<double> &B,
+         matrix::Dense<double> &C);
+void sub(const matrix::Dense<double> &A, const matrix::Dense<double> &B,
+         matrix::Dense<double> &C);
+void mul(const matrix::Dense<double> &A, const matrix::Dense<double> &B,
+         matrix::Dense<double> &C);
+void div(const matrix::Dense<double> &A, const matrix::Dense<double> &B,
+         matrix::Dense<double> &C);
+
+void add(const matrix::Dense<double> &A, const double alpha,
+         matrix::Dense<double> &C);
+void sub(const matrix::Dense<double> &A, const double alpha,
+         matrix::Dense<double> &C);
+void mul(const matrix::Dense<double> &A, const double alpha,
+         matrix::Dense<double> &C);
+void div(const matrix::Dense<double> &A, const double alpha,
+         matrix::Dense<double> &C);
+
+void add(const matrix::CRS<double> &A, const matrix::CRS<double> &B,
+         matrix::CRS<double> &C);
+void sub(const matrix::CRS<double> &A, const matrix::CRS<double> &B,
+         matrix::CRS<double> &C);
+void mul(const matrix::CRS<double> &A, const matrix::CRS<double> &B,
+         matrix::CRS<double> &C);
+void div(const matrix::CRS<double> &A, const matrix::CRS<double> &B,
+         matrix::CRS<double> &C);
+
+void add(const matrix::CRS<double> &A, const double alpha,
+         matrix::CRS<double> &C);
+void sub(const matrix::CRS<double> &A, const double alpha,
+         matrix::CRS<double> &C);
+void mul(const matrix::CRS<double> &A, const double alpha,
+         matrix::CRS<double> &C);
+void div(const matrix::CRS<double> &A, const double alpha,
+         matrix::CRS<double> &C);
 //////////////////////////////////////////////////////
 //  Vector
 //////////////////////////////////////////////////////