+++ src/blas/matrix/matadd/d2f.sh
@@ -1,3 +1,0 @@
-cat double_matadd.cpp \
-    | sed -e 's/double/float/g' \
-    > float_matadd.cpp