+++ src/blas/matrix/mataddsub/d2f.sh
@@ -1,0 +1,3 @@
+cat double_mataddsub.cpp \
+    | sed -e 's/double/float/g' \
+    > float_mataddsub.cpp