+++ CHANGELOG.md
@@ -21,11 +21,14 @@ https://keepachangelog.com/ja/1.0.0/ に基づいて記述していく
 Unreleased
 -----------
 ### Added
+- matrix四則演算関数を追加 https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/149
+- matadd/matsub関数を追加 https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/149
 - vecadd/vecsubを追加 https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/148
 - vector四則演算関数を追加 https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/147
 - matrix subを追加 https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/146
 
 ### Changed 
+- matadd/をmataddsub/に変更 https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/149
 - test/benchmarkのvector_commonをoperatorでなく四則演算関数に変更 https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/147
 - matrix add/subでdoubleからfloatを作るようにファイル構成を変更  https://gitlab.ritc.jp/ricos/monolish/-/merge_requests/146